Moved base class to base.py
